+++ packages/_vue3-migration-test/src/main.ts
@@ -1,19 +1,29 @@
 import { QuerySuggestionsRequest, XComponentsAdapter } from '@empathyco/x-types';
 import { Component, configureCompat, createApp } from 'vue';
 import { createStore } from 'vuex';
-import { xPlugin } from '../../x-components/src/plugins/x-plugin';
 import { getRelatedTagsStub } from '../../x-components/src/__stubs__/related-tags-stubs.factory';
 import { getQuerySuggestionsStub } from '../../x-components/src/__stubs__/query-suggestions-stubs.factory';
+import {
+  createResultStub,
+  createRedirectionStub,
+  getBannersStub,
+  getNextQueriesStub,
+  getPromotedsStub,
+  getResultsStub
+} from '../../x-components/src/__stubs__/index';
+import { XInstaller } from '../../x-components/src/x-installer/x-installer/x-installer';
 import App from './App.vue';
 import router from './router';
 import {
   facetsXModule,
   nextQueriesXModule,
+  popularSearchesXModule,
   queriesPreviewXModule,
   recommendationsXModule,
   scrollXModule,
   searchXModule,
-  semanticQueriesXModule
+  semanticQueriesXModule,
+  identifierResultsXModule
 } from './';
 
 // Warnings that cannot be solved in Vue 2 (a.k.a. breaking  changes) are suppressed
@@ -31,8 +41,7 @@ if (VUE_COMPAT_MODE === 2) {
     INSTANCE_LISTENERS: 'suppress-warning',
     INSTANCE_ATTRS_CLASS_STYLE: 'suppress-warning',
     RENDER_FUNCTION: false,
-    COMPONENT_V_MODEL: false,
-    WATCH_ARRAY: false
+    COMPONENT_V_MODEL: false
   });
 }
 
@@ -44,7 +53,37 @@ const adapter = {
   querySuggestions: (request: QuerySuggestionsRequest) =>
     new Promise(resolve => {
       resolve({ suggestions: getQuerySuggestionsStub(request.query, 5) });
-    })
+    }),
+  nextQueries: () =>
+    new Promise(resolve => {
+      resolve({
+        nextQueries: [
+          ...getNextQueriesStub(),
+          {
+            facets: [],
+            isCurated: false,
+            modelName: 'NextQuery',
+            query: 'next_query_preview',
+            results: getResultsStub(10),
+            totalResults: 10
+          }
+        ]
+      });
+    }),
+  search: () =>
+    new Promise(resolve => {
+      resolve({
+        results: getResultsStub(10),
+        totalResults: 50,
+        promoteds: getPromotedsStub(),
+        banners: getBannersStub(),
+        redirections: [createRedirectionStub('redirection')]
+      });
+    }),
+  identifierResults: () =>
+    new Promise(resolve =>
+      resolve({ results: ['123A', '123B', '123C', '123D'].map(id => createResultStub(id)) })
+    )
 } as unknown as XComponentsAdapter;
 
 const store = createStore({});
@@ -52,17 +91,24 @@ const store = createStore({});
 createApp(App as Component)
   .use(router)
   .use(store)
-  .use(xPlugin, {
-    adapter,
-    store,
-    __PRIVATE__xModules: {
-      facets: facetsXModule,
-      nextQueries: nextQueriesXModule,
-      scroll: scrollXModule,
-      search: searchXModule,
-      queriesPreview: queriesPreviewXModule,
-      semanticQueries: semanticQueriesXModule,
-      recommendations: recommendationsXModule
-    }
-  })
   .mount('#app');
+
+window.initX = {
+  instance: 'empathy',
+  lang: 'en'
+};
+new XInstaller({
+  adapter,
+  store,
+  __PRIVATE__xModules: {
+    facets: facetsXModule,
+    nextQueries: nextQueriesXModule,
+    scroll: scrollXModule,
+    search: searchXModule,
+    queriesPreview: queriesPreviewXModule,
+    semanticQueries: semanticQueriesXModule,
+    recommendations: recommendationsXModule,
+    identifierResults: identifierResultsXModule,
+    popularSearches: popularSearchesXModule
+  }
+}).init();