+++ packages/_vue3-migration-test/src/x-modules/identifier-results/index.ts
@@ -1,2 +1,2 @@
-export * from './x-module';
 export * from './components';
+export * from './x-module';